Move PhotoController to 'controller' package
